requirements.txt

streamlit==1.42.0
google-genai==1.3.0
Pillow==10.2.0
